rename: privateClaims 패키지 이동
