+++ backend/ddang/src/main/java/com/ddang/ddang/authentication/application/AuthenticationService.java
@@ -3,7 +3,7 @@ package com.ddang.ddang.authentication.application;
 import com.ddang.ddang.authentication.application.dto.TokenDto;
 import com.ddang.ddang.authentication.domain.exception.InvalidTokenException;
 import com.ddang.ddang.authentication.domain.Oauth2UserInformationProviderComposite;
-import com.ddang.ddang.authentication.domain.PrivateClaims;
+import com.ddang.ddang.authentication.infrastructure.jwt.PrivateClaims;
 import com.ddang.ddang.authentication.domain.TokenDecoder;
 import com.ddang.ddang.authentication.domain.TokenEncoder;
 import com.ddang.ddang.authentication.domain.TokenType;
