+++ backend/ddang/src/main/java/com/ddang/ddang/authentication/configuration/AuthenticationInterceptor.java
@@ -1,6 +1,6 @@
 package com.ddang.ddang.authentication.configuration;
 
-import com.ddang.ddang.authentication.domain.PrivateClaims;
+import com.ddang.ddang.authentication.infrastructure.jwt.PrivateClaims;
 import com.ddang.ddang.authentication.domain.TokenDecoder;
 import com.ddang.ddang.authentication.domain.TokenType;
 import com.ddang.ddang.authentication.domain.dto.AuthenticationStore;
