--- backend/ddang/src/main/java/com/ddang/ddang/authentication/domain/PrivateClaims.java
@@ -1,4 +1,0 @@
-package com.ddang.ddang.authentication.domain;
-
-public record PrivateClaims(Long userId) {
-}
