+++ backend/ddang/src/main/java/com/ddang/ddang/authentication/domain/TokenDecoder.java
@@ -1,5 +1,6 @@
 package com.ddang.ddang.authentication.domain;
 
+import com.ddang.ddang.authentication.infrastructure.jwt.PrivateClaims;
 import java.util.Optional;
 
 public interface TokenDecoder {
