+++ backend/ddang/src/main/java/com/ddang/ddang/authentication/infrastructure/jwt/PrivateClaims.java
@@ -1,0 +1,4 @@
+package com.ddang.ddang.authentication.infrastructure.jwt;
+
+public record PrivateClaims(Long userId) {
+}
